support favicon display by using a pre-built collection
Add support to load favicons from a common URI based on the
hostname of the bookmark as it is nearly impossible to guess the
favicon of the bookmark correctly.

As an example the DuckDuckGo collection can be used, but it is
a small one and a bit unfair to us their bandwidth (beside that
requesting a bunch of small files from remote isn't exactly fast, too)
So you really want to roll your own collection if you want to use
this feature. Otherwise just leave it disabled, which is the default.

[Forward-port of 60da94895e8d0725ac72e335898b9ef20fc8b856
 from Sun Apr 21 17:11:43 2013 +0200]